requirements.txt
groq
python-dotenv
pypdf
gradio
requests